Rename game folder to js to be more consistent
