+++ static/scenes/railroad/js/camera-system.js
